refactor: add service adapter module
